[#111-add-socket-client] Move server tests into tests/<type>/server/
